[cmake] changed how external deps built
An external directory is added. The cmake list file in that directory
gathers the build of all external directories and consider also
veriparse as an external project.

Thus veriparse can use the native find_package mechanism to find
libraries.

This will ease to setup the choice between building an external library
or using the version found in the system.