[tests] fixed mpz_class initialization with a literal
By default, literals are signed. When the value exceed 2^31, a int64_t
is used. But on windows, int64_t are translated into long long
int. mpz_class cannot be initialized with a long long (cf man).

By putting a 'u' a the end of the literal, we will just use an uint32_t
type, and everything will be fine on both linux and windows.